address PR #31 review
- Split `src/commands/install/render.ts` (216 lines) into a
  directory. `render/index.ts` is 147 lines (renderInstall +
  the two block renderers); `render/helpers.ts` is 85 lines
  (renderAgentLine, tallyAgents, formatTotals, formatVersion).
  Both well under the 200-line cap.
- Align `renderAlreadyInstalled`'s project-scope tag with the
  shape used by `renderRecord`: two leading spaces inside
  `style.dim` (previously had the leading space outside). Purely
  cosmetic consistency.
- Tighten the C-INST-11 regex from `/claude-code .+demo/` to
  `/[-·]\s+claude-code\s+→.+demo/` so a regression in the
  muted-marker or the "→" separator now surfaces here.
