+++ src/commands/install/index.ts
@@ -14,7 +14,7 @@ import { runInstall } from "../../install/flow.ts";
 import type { KindHint } from "../../install/resolve-ref/index.ts";
 import type { CommandContext, CommandOutput } from "../types.ts";
 import { promptBareNameAmbiguity } from "./ambiguity-prompt.ts";
-import { renderInstall } from "./render.ts";
+import { renderInstall } from "./render/index.ts";
 
 /** Read --tap / --bundle / --skill, enforce mutual exclusivity. */
 function readKindHint(ctx: CommandContext): KindHint {
